# Conflicts:
#	plugins/txes/storage/src/plugins/DataModificationApprovalTransactionPlugin.cpp
#	plugins/txes/storage/src/plugins/DownloadApprovalTransactionPlugin.cpp
#	plugins/txes/storage/src/validators/DataModificationApprovalUploadWorkValdidator.cpp
#	plugins/txes/storage/tests/validators/DownloadApprovalValidatorTests.cpp
#	src/catapult/model/StorageNotifications.h

+++ plugins/txes/storage/tests/test/StorageTestUtils.h
@@ -200,7 +200,7 @@ namespace catapult { namespace test {
 		uint8_t JudgingKeysCount;
 		uint8_t OverlappingKeysCount;
 		uint8_t JudgedKeysCount;
-		uint8_t OpinionElementCount;
+		uint16_t OpinionElementCount;
 		int16_t FilledPresenceRowIndex;
 		std::vector<Key> PublicKeys;
 		std::vector<Signature> Signatures;
